backend/requirements.txt
# AI Blog Assistant Backend Requirements

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML APIs
openai==1.3.7
anthropic==0.7.8

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Email
fastapi-mail==1.4.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Database drivers
# PostgreSQL
psycopg2-binary==2.9.9

# MySQL
# pymysql==1.1.0

# Monitoring (optional)
# sentry-sdk[fastapi]==1.38.0